Fix broken entry processor documentation link (#221)
[Slack
discussion](https://hazelcast.slack.com/archives/C035HQET5/p1722580449491389)

Note - I am *unsure* what the URL is used for to identify the problem
with the current behaviour, I _think_ it's used to redirect old links to
the "new" location, but I'm not sure, so I am unable to test this
change. But the current URLs no longer exist.